deleted:    Database Model modified:   assets/css/estilos.css new file:   assets/img/btn-addT.png new file:   assets/img/editar.png new file:   assets/img/eliminar.png new file:   assets/img/form-trans.png deleted:    assets/img/logo1.png new file:   assets/img/name.png new file:   assets/img/rif.png new file:   assets/img/screen-home.png new file:   assets/img/screen-login.png new file:   assets/img/screen-newpass.png new file:   assets/img/screen-recover.png deleted:    assets/img/welcome.jpeg new file:   assets/img/widget-chart.png new file:   assets/img/widget-cuentas.png new file:   assets/img/widget-datos1.png new file:   assets/img/widget-datos2.png new file:   assets/img/widget-datos3.png new file:   assets/img/widget-datos4.png new file:   assets/img/widget-historial.png new file:   assets/img/widget-menu.png new file:   assets/img/widget-perfil-ad.png new file:   assets/img/widget-perfil.png new file:   assets/img/widget-report-btn.png new file:   assets/img/widget-report-form.png new file:   assets/img/widget-tablaT.png new file:   assets/img/widget-ubicaciones-info.png new file:   assets/img/widget-ubicaciones-list.png modified:   assets/js/ajax/principal.js modified:   conexiones/create.php modified:   conexiones/funciones.php modified:   conexiones/historial.php modified:   conexiones/inventarioUpdate.php modified:   conexiones/login.php modified:   conexiones/logout.php new file:   conexiones/report.php modified:   conexiones/reportes.php modified:   conexiones/ubicaciones.php new file:   controllers/ayuda.php new file:   controllers/configuraciones.php new file:   controllers/transformador.php modified:   modulos/header.php new file:   sistema-corpoelec.sql new file:   views/ayuda/ayuda.php deleted:    views/busqueda/busqueda.php new file:   views/configuraciones/configuraciones.php deleted:    views/create/create.php modified:   views/dashboard/dashboard.php modified:   views/delete/delete.php modified:   views/editar/editar.php modified:   views/historial/historial.php modified:   views/home.php modified:   views/inventario/inventario.php modified:   views/login/login.php modified:   views/newPass/newPass.php modified:   views/recover/recover.php modified:   views/reportes/reportes.php new file:   views/transformador/transformador.php modified:   views/ubicaciones/ubicaciones.php deleted:    views/userDashboard/userDashboard.php

+++ views/historial/historial.php
@@ -45,7 +45,7 @@
   <div class="container-fluid p-4">
     <div id="accordion-one" class="accordion">
       <div class="d-flex flex-row justify-content-space ml-5">
-        <button class="mb-0 btn btn-primary mx-1" data-toggle="collapse" data-target="#collapseOne" aria-expanded="true" aria-controls="collapseOne"><i class="bx bx-plus-circle"></i> Añadir Operación</button>
+        <button class="mb-0 btn btn-primary mx-1" data-toggle="collapse" data-target="#collapseOne" aria-expanded="true" aria-controls="collapseOne"><i class="bx bx-plus-circle text-white"></i> Añadir Operación</button>
       </div>
       
       <div id="collapseOne" class="collapse card mt-3 col-9 rounded mx-auto" data-parent="#accordion-one">
@@ -96,7 +96,11 @@
                 <th>Fecha</th>
                 <th>N° Serial del Transformador</th>
                 <th>Estado</th>
-                <th>Acciones</th>
+                <?php
+                  if($_SESSION['tipo'] == "Administrador") {
+                    echo '<th>Acciones</th>';
+                  }
+                ?>
               </tr>
             </thead>
             <tbody>
@@ -110,18 +114,22 @@
                         <th> <strong>" . $num++ . "</strong></th>
                         <td>" . $rows['O_Procedimiento'] . "</td>
                         <td>" . $rows['O_Fecha'] . "</td>
-                        <td><a class='text-info' href='equipo?serial=" . $rows['O_Equipo'] . "'>" . $rows['O_Equipo'] . "</a></td>
+                        <td><a class='text-info' href='transformador?serial=" . $rows['O_Equipo'] . "'>" . $rows['O_Equipo'] . "</a></td>
                         <td>" . $rows['O_EstadoActual'] . "</td>
-                        <td class='mt-0'>
+                        ";
+
+                        if($_SESSION['tipo'] == "Administrador") {
+                        echo "<td class='mt-0'>
                           <a class='btn btn-sm btn-info' href='editar?operacion=" . $rows['O_Codigo'] . "'>
-                            <span class='tf-icons bx bx-edit'></span>
+                            <span class='tf-icons bx bx-edit text-white'></span>
                           </a>
 
                           <a class='btn btn-sm btn-danger' href='delete?operacion=" . $rows['O_Codigo'] . "'>
-                            <span class='tf-icons bx bx-trash'></span>
+                            <span class='tf-icons bx bx-trash text-white'></span>
                           </a>
-                        </td>
-                      </tr>";
+                        </td>";
+                        }
+                      echo "</tr>";
                 };  
               ?>
             </tbody>
